# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List cleaning jobs
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - name: status
          in: query
          schema:
            type: string
            enum: [scheduled, in_progress, completed, attention]
      responses:
        "200":
          description: Cleaning jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
    post:
      operationId: createJob
      tags: [jobs]
      summary: Create a cleaning job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInput"
      responses:
        "201":
          description: Created job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get a cleaning job
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Cleaning job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Job not found
    patch:
      operationId: updateJob
      tags: [jobs]
      summary: Update a cleaning job
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobUpdate"
      responses:
        "200":
          description: Updated job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{id}/checklist:
    patch:
      operationId: updateJobChecklist
      tags: [jobs]
      summary: Update checklist completion for a job
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChecklistUpdate"
      responses:
        "200":
          description: Updated job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{id}/messages:
    post:
      operationId: sendJobMessage
      tags: [jobs]
      summary: Send a job message
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Sent message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /team:
    get:
      operationId: listTeam
      tags: [team]
      summary: List team members
      responses:
        "200":
          description: Team members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeamMember"
    post:
      operationId: createTeamMember
      tags: [team]
      summary: Add a team member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamMemberInput"
      responses:
        "201":
          description: Created team member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"
  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [storage]
      summary: Request a protected direct-to-storage upload URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadInput"
      responses:
        "201":
          description: Upload URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrl"
...
  /jobs/{jobId}/messages:
    get:
      operationId: listJobMessages
      tags: [field-ops]
      parameters: [{ $ref: "#/components/parameters/JobId" }]
      responses: { "200": { description: Messages, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/MessageRecord" } } } } } }
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
    JobId:
      name: jobId
      in: path
      required: true
      schema: { type: integer }
  schemas:
...
    Job:
      type: object
      required: [id, clientName, address, scheduledDate, startTime, endTime, status, serviceType, team, checklist]
      properties:
        id:
          type: integer
        clientName:
          type: string
        address:
          type: string
        scheduledDate:
          type: string
          format: date
        startTime:
          type: string
        endTime:
          type: string
        status:
          type: string
          enum: [scheduled, in_progress, completed, attention]
        serviceType:
          type: string
        serviceVariant:
          type: ["string", "null"]
        addOns:
          type: array
          items:
            type: string
        durationMinutes:
          type: ["integer", "null"]
        frequency:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        team:
          type: array
          items:
            $ref: "#/components/schemas/TeamMember"
        checklist:
          type: array
          items:
            $ref: "#/components/schemas/ChecklistItem"
        photos:
          type: array
          items:
            $ref: "#/components/schemas/Photo"
        clientPhone:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    JobInput:
      type: object
      required: [clientName, address, scheduledDate, startTime, endTime, serviceType]
      properties:
        clientName:
          type: string
          minLength: 1
        address:
          type: string
          minLength: 1
        scheduledDate:
          type: string
          format: date
        startTime:
          type: string
        endTime:
          type: string
        serviceType:
          type: string
        serviceVariant:
          type: string
        addOns:
          type: array
          items:
            type: string
        durationMinutes:
          type: integer
          minimum: 0
        frequency:
          type: string
        notes:
          type: string
        clientPhone:
          type: string
        teamMemberIds:
          type: array
          items:
            type: integer
    JobUpdate:
      type: object
      properties:
        scheduledDate:
          type: string
          format: date
        startTime:
          type: string
        endTime:
          type: string
        status:
          type: string
          enum: [scheduled, in_progress, completed, attention]
        notes:
          type: string
        serviceVariant:
          type: string
        addOns:
          type: array
          items:
            type: string
        durationMinutes:
          type: integer
          minimum: 0
        frequency:
          type: string
        teamMemberIds:
          type: array
          items:
            type: integer
    ChecklistItem:
      type: object
      required: [id, label, completed]
      properties:
        id:
          type: integer
        label:
          type: string
        completed:
          type: boolean
    ChecklistUpdate:
      type: object
      required: [itemId, completed]
      properties:
        itemId:
          type: integer
        completed:
          type: boolean
    Photo:
      type: object
      required: [id, url, label, createdAt]
      properties:
        id:
          type: integer
        url:
          type: string
          format: uri
        label:
          type: string
        createdAt:
          type: string
          format: date-time
    TeamMember:
      type: object
      required: [id, name, role, phone, status]
      properties:
        id:
          type: integer
        name:
          type: string
        role:
          type: string
        phone:
          type: string
        status:
          type: string
          enum: [available, assigned, off]
        initials:
          type: string
    TeamMemberInput:
      type: object
      required: [name, role, phone]
      properties:
        name:
          type: string
        role:
          type: string
        phone:
          type: string
    MessageInput:
      type: object
      required: [recipient, body]
      properties:
        recipient:
          type: string
          enum: [client, team]
        body:
          type: string
          minLength: 1
        channel:
          type: string
          enum: [sms, whatsapp]
          default: sms
        audience:
          type: string
          enum: [customer, employee]
          default: customer
        recipientPhone:
          type: string
        recipientName:
          type: string
        metadata:
          type: object
    Message:
      type: object
      required: [id, recipient, body, channel, audience, status, createdAt]
      properties:
        id:
          type: integer
        recipient:
          type: string
          enum: [client, team]
        body:
          type: string
        channel:
          type: string
          enum: [sms, whatsapp]
        audience:
          type: string
          enum: [customer, employee]
        recipientPhone:
          type: ["string", "null"]
        recipientName:
          type: ["string", "null"]
        providerMessageId:
          type: ["string", "null"]
        status:
          type: string
          enum: [queued, sent, failed]
        createdAt:
          type: string
          format: date-time
        sentAt:
          type: ["string", "null"]
          format: date-time
        deliveredAt:
          type: ["string", "null"]
          format: date-time
        failedAt:
          type: ["string", "null"]
          format: date-time
        failureReason:
          type: ["string", "null"]
        metadata:
          type: object
        actorClerkUserId:
          type: ["string", "null"]
    UploadInput:
      type: object
      required: [name, size, contentType]
      properties:
        name: { type: string, minLength: 1 }
        size: { type: integer, minimum: 1 }
        contentType: { type: string, minLength: 1 }
    UploadUrl:
      type: object
      required: [uploadURL, objectPath, metadata]
      properties:
        uploadURL: { type: string, format: uri }
        objectPath: { type: string }
        metadata:
          type: object
          required: [name, size, contentType]
          properties:
            name: { type: string }
            size: { type: integer }
            contentType: { type: string }
...
    MessageRecord:
      type: object
      required: [id, recipient, body, channel, audience, status, provider, createdAt]
      properties:
        id: { type: integer }
        recipient: { type: string }
        body: { type: string }
        channel: { type: string, enum: [sms, whatsapp] }
        audience: { type: string, enum: [customer, employee] }
        status: { type: string }
        provider: { type: string }
        recipientPhone: { type: ["string", "null"] }
        recipientName: { type: ["string", "null"] }
        providerMessageId: { type: ["string", "null"] }
        sentAt: { type: ["string", "null"], format: date-time }
        deliveredAt: { type: ["string", "null"], format: date-time }
        failedAt: { type: ["string", "null"], format: date-time }
        failureReason: { type: ["string", "null"] }
        metadata: { type: object }
        actorClerkUserId: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }